Add LD_PRELOAD version of groot
This adds a version of the "groot" commandline that works as a
LD_PRELOAD library. If you add libgroot.so to LD_PRELOAD and set
GROOT_ENABLED=1, then all processes will spawn in its own namespace
similar to thow running something in the groot CLI works.

This is useful for using groot in yocto, as we can enable groot
automatically for all spawned processes.