preload: Make unsetenv work in the init
For some weird reason I can't understand unsetenv() doesn't seem to
work when called in the init function, so I replaced the libc call it
with a trivial reimplementation.

You could see this by using something like:

LD_PRELOAD=`pwd`/libgroot.so /bin/sh -c 'echo a | cat | cat '
Which will print:
  groot: Warning: no defined ids for user root in /etc/subuid, limited user/group support
  groot: Warning: no defined ids for user root in /etc/subgid, limited user/group support
  groot: Warning: no defined ids for user root in /etc/subuid, limited user/group support
  groot: Warning: no defined ids for user root in /etc/subgid, limited user/group support
a

Because the echo and the cat processes also try to do the LD_PRELOAD.

We also set a GROOT_DISABLED flag to combat reetrancy because in yocto sometimes the LD_PRELOAD
is re-added to ten environment, even though we unset it.

foo